Syntax highlighter (hljs), jpeg quality (upload) upped

+++ MongSpeak/upload.h
@@ -55,7 +55,7 @@ size_t  FileReadCb(void*  _Buffer, size_t _ElementSize, size_t _ElementCount, up
 	size_t tosend = fread(_Buffer, _ElementSize, _ElementCount, _Stream->fd);
 	_Stream->sent_bt += tosend * _ElementSize;
 	if (g_network)
-		g_network->nBytesWritten += tosend * _ElementSize;
+		g_network->nBytesWritten += (int)(tosend * _ElementSize);
 	int uploaded_pc = _Stream->sent_bt / _Stream->size * 100;
 	if (uploaded_pc > _Stream->sent_pc) {
 		_Stream->sent_pc = uploaded_pc;
